+++ src/client.ts
@@ -1,2 +1,2 @@
 export type * from './types.ts'
-
+export { buildDiagnosticReport } from './report.ts'